CADC-9555 - complete the move of client-based functions from manage_composable.py to client_composable.py.
