+++ caom2pipe/tests/test_caom_composable.py
@@ -181,7 +181,7 @@ def test_reset():
     assert len(all_archive_keys) == 2, 'wrong artifact count'
 
 
-@patch('caom2pipe.manage_composable.query_tap_client')
+@patch('caom2pipe.client_composable.query_tap_client')
 def test_build_temporal_wcs(query_mock):
     def _mock_query(arg1, arg2):
         if 'N20160102S0296' in arg1:
